requirements.txt
django
django-cors-headers
djangorestframework
mysqlclient
boto3
Werkzeug